app/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.7.2
    # via
    #   -r requirements.in
    #   django
async-timeout==4.0.3
    # via
    #   -r requirements.in
    #   redis
attrs==23.2.0
    # via
    #   -r requirements.in
    #   jsonschema
    #   referencing
build==1.0.3
    # via
    #   -r requirements.in
    #   pip-tools
certifi==2023.11.17
    # via
    #   -r requirements.in
    #   requests
charset-normalizer==3.3.2
    # via
    #   -r requirements.in
    #   requests
click==8.1.7
    # via
    #   -r requirements.in
    #   pip-tools
django==5.0.1
    # via
    #   -r requirements.in
    #   django-debug-toolbar
    #   django-redis
    #   djangorestframework
    #   drf-spectacular
django-debug-toolbar==4.2.0
    # via -r requirements.in
django-redis==5.4.0
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   drf-spectacular
drf-spectacular==0.27.1
    # via -r requirements.in
gunicorn==21.2.0
    # via -r requirements.in
idna==3.6
    # via
    #   -r requirements.in
    #   requests
inflection==0.5.1
    # via
    #   -r requirements.in
    #   drf-spectacular
jsonschema==4.21.1
    # via
    #   -r requirements.in
    #   drf-spectacular
jsonschema-specifications==2023.12.1
    # via
    #   -r requirements.in
    #   jsonschema
packaging==23.2
    # via
    #   -r requirements.in
    #   build
    #   gunicorn
pillow==10.2.0
    # via -r requirements.in
pip-tools==7.3.0
    # via -r requirements.in
psycopg2==2.9.9
    # via -r requirements.in
pyproject-hooks==1.0.0
    # via
    #   -r requirements.in
    #   build
python-decouple==3.8
    # via -r requirements.in
pytz==2023.4
    # via
    #   -r requirements.in
    #   djangorestframework
pyyaml==6.0.1
    # via
    #   -r requirements.in
    #   drf-spectacular
redis==5.0.1
    # via
    #   -r requirements.in
    #   django-redis
referencing==0.33.0
    # via
    #   -r requirements.in
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via -r requirements.in
rpds-py==0.17.1
    # via
    #   -r requirements.in
    #   jsonschema
    #   referencing
sqlparse==0.4.4
    # via
    #   -r requirements.in
    #   django
    #   django-debug-toolbar
tomli==2.0.1
    # via
    #   -r requirements.in
    #   build
    #   pip-tools
    #   pyproject-hooks
typing-extensions==4.9.0
    # via
    #   -r requirements.in
    #   asgiref
uritemplate==4.1.1
    # via
    #   -r requirements.in
    #   drf-spectacular
urllib3==2.2.0
    # via
    #   -r requirements.in
    #   requests
wheel==0.42.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
